JSON-C removal

+++ module/crypto/src/dap_cert_file.c
@@ -275,9 +275,9 @@ uint8_t* dap_cert_mem_save(dap_cert_t * a_cert, uint32_t *a_cert_size_out)
 {
     dap_enc_key_t *l_key = a_cert->enc_key;
 
-    uint64_t  l_priv_key_data_size = a_cert->enc_key->key_pvt_size,
-            l_pub_key_data_size = a_cert->enc_key->pub_key_size,
-            l_metadata_size = l_key->_inheritor_size;
+    size_t l_priv_key_data_size = a_cert->enc_key->key_pvt_size,
+           l_pub_key_data_size = a_cert->enc_key->pub_key_size,
+           l_metadata_size = l_key->_inheritor_size;
             
     uint8_t *l_pub_key_data = dap_enc_key_serialize_pub_key(l_key, &l_pub_key_data_size),
             *l_priv_key_data = (l_key->priv_key_data && l_key->priv_key_data_size) ?
